python/requirements.txt
# MirrorWitness 2025-11-04

websockets==12.0
aiohttp==3.9.1
PyNaCl==1.5.0
